+++ deploy_prod.sh
@@ -1,8 +1,17 @@
-# npm version patch
+# npm version prerelease --preid=beta
 version=`node -e 'console.log(require("./package.json").version)'`
-ECHO "____________WIDGET-V5______________"
+echo "version $version"
+echo "____________WIDGET-V5______________"
 echo "CREATING TAG ON GIT FOR version: $version"
-# echo "version $version"
+
+# Get curent branch name
+current_branch=$(git rev-parse --abbrev-ref HEAD)
+remote_name=$(git config --get branch.$current_branch.remote)
+
+## Push commit to git
+git add .
+git commit -m "version added: ### $version"
+git push "$remote_name" "$current_branch"
 
 # Get curent branch name
 current_branch=$(git rev-parse --abbrev-ref HEAD)