+++ src/app/app.component.ts
@@ -4,6 +4,7 @@ import { AfterViewInit, Component, ElementRef, HostListener, NgZone, OnDestroy, 
 import { environment } from 'src/environments/environment';
 import { Subscription } from 'rxjs/internal/Subscription';
 import { v4 as uuidv4 } from 'uuid';
+import { HEADER_MENU_OPTION } from './utils/constants';
 //COMPONENTS
 import { EyeeyeCatcherCardComponent } from './component/eyeeye-catcher-card/eyeeye-catcher-card.component';
 //MODELS
@@ -100,6 +101,8 @@ export class AppComponent implements OnInit, AfterViewInit, OnDestroy {
   
   @ViewChild(EyeeyeCatcherCardComponent, { static: false }) eyeeyeCatcherCardComponent: EyeeyeCatcherCardComponent
   styleMapConversation: Map<string, string> = new Map();
+  translationMap: Map<string, string> = new Map();
+  isButtonsDisabled: boolean = true;
   marginBottom: number;
   
   forceDisconnect: boolean = false;
@@ -139,6 +142,12 @@ export class AppComponent implements OnInit, AfterViewInit, OnDestroy {
 
     ngAfterViewInit(): void {
         this.logger.info('[APP-CONF]---------------- ngAfterViewInit: APP.COMPONENT ---------------- ')
+        
+        // Initialize translation map and enable buttons
+        const keys = ['MAXIMIZE', 'MINIMIZE', 'CENTER', 'BUTTON_CLOSE_TO_ICON'];
+        this.translationMap = this.translateService.translateLanguage(keys);
+        this.isButtonsDisabled = false;
+        
         this.ngZone.run(() => {
             const that = this;
             const subAddedConversation = this.conversationsHandlerService.conversationAdded.subscribe((conversation) => {
@@ -262,6 +271,14 @@ export class AppComponent implements OnInit, AfterViewInit, OnDestroy {
                 this.tabTitle = this.g.windowContext.window.document.title
                 this.appStorageService.initialize(environment.storage_prefix, this.g.persistence, this.g.projectid)
                 
+                //check if allowed to load
+                let canLoad = this.globalSettingsService.manageLoadingDomains();
+                if(!canLoad){
+                    console.error('[Check canLoad] Widget is not able to load on this domain!!!')
+                    this.hideWidget()
+                    this.disposeWidget();
+                }
+
                 //set visibility
                 if((this.g.isMobile && !this.g.displayOnMobile) || (!this.g.isMobile && !this.g.displayOnDesktop)){
                     this.disposeWidget()
@@ -1980,6 +1997,39 @@ export class AppComponent implements OnInit, AfterViewInit, OnDestroy {
         this.signOut();
     }
 
+    onMenuOptionClick(event: string) {
+        this.logger.debug('[APP-COMP] onMenuOptionClick', event);
+        switch(event) {
+            case HEADER_MENU_OPTION.MAXIMIZE:
+                this.onWidgetSizeChange('max');
+                break;
+            case HEADER_MENU_OPTION.MINIMIZE:
+                this.onWidgetSizeChange('min');
+                break;
+            case HEADER_MENU_OPTION.TOP:
+                this.onWidgetSizeChange('top');
+                break;
+        }
+    }
+
+    onWidgetSizeChange(mode: 'min' | 'max' | 'top') {
+        var tiledeskDiv = this.g.windowContext.window.document.getElementById('tiledeskdiv') 
+        this.g.size = mode 
+        if(mode==='max'){
+            tiledeskDiv.classList.add('max-size')
+            tiledeskDiv.classList.remove('min-size')
+            tiledeskDiv.classList.remove('top-size')
+        }else if(mode==='min'){
+            tiledeskDiv.classList.add('min-size')
+            tiledeskDiv.classList.remove('max-size')
+            tiledeskDiv.classList.remove('top-size')
+        }else if(mode=== 'top'){
+            tiledeskDiv.classList.add('top-size')
+            tiledeskDiv.classList.remove('max-size')
+            tiledeskDiv.classList.remove('min-size')
+        }
+    }
+
     /**
      * MODAL RATING WIDGET:
      * close modal page
@@ -2110,6 +2160,13 @@ export class AppComponent implements OnInit, AfterViewInit, OnDestroy {
 
         this.el.nativeElement.style.setProperty('--button-in-msg-background-color', this.g.bubbleSentBackground)
         this.el.nativeElement.style.setProperty('--button-in-msg-font-size', this.g.buttonFontSize)
+        this.el.nativeElement.style.setProperty('--button-in-msg-font-family', this.g.fontFamily)
+
+        this.el.nativeElement.style.setProperty('--chat-header-height', this.g.hideHeaderConversation? '0px': null)
+        this.el.nativeElement.style.setProperty('--font-size-bubble-message', this.g.fontSize)
+        this.el.nativeElement.style.setProperty('--font-family-bubble-message', this.g.fontFamily)
+        
+
     }