+++ src/app/component/conversation-detail/conversation-header/conversation-header.component.spec.ts
@@ -1,8 +1,8 @@
-import { HttpClientModule } from '@angular/common/http';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import { AppConfigService } from './../../../providers/app-config.service';
 import { Globals } from './../../../utils/globals';
 import { NO_ERRORS_SCHEMA } from '@angular/core';
-import { async, ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { ConversationHeaderComponent } from './conversation-header.component';
 import { TypingService } from '../../../../chat21-core/providers/abstract/typing.service';
@@ -18,17 +18,16 @@ describe('ConversationHeaderComponent', () => {
   
   beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
-      declarations: [ ConversationHeaderComponent ],
-      imports: [
-        HttpClientModule
-      ],
-      providers: [ 
+    declarations: [ConversationHeaderComponent],
+    schemas: [NO_ERRORS_SCHEMA],
+    imports: [],
+    providers: [
         Globals,
         TypingService,
-        AppConfigService
-      ],
-      schemas: [NO_ERRORS_SCHEMA],
-    })
+        AppConfigService,
+        provideHttpClient(withInterceptorsFromDi())
+    ]
+})
     .compileComponents();
   }));